Refactor AidsPanel
